backend/requirements.txt
blurhash==1.1.4
fastapi==0.115.6
imagehash==4.3.1
numpy==2.2.2
passlib==1.7.4
Pillow==11.1.0
psutil==6.1.1
pydantic==2.10.5
PyJWT==2.10.1
pymilvus==2.5.3
python-dotenv==1.0.1
python_magic==0.4.27
SQLAlchemy==2.0.37
transformers==4.48.0
uvicorn
python-multipart
psycopg2-binary
minio